requirements.txt
flask
pandas
python-docx
PyPDF2
